requirements.txt
PyPDF2
transformers
pymupdf
sentencepiece
protobuf
python-dotenv